pyproject.toml
[tool.poetry]
name = "tidalpresence"
version = "0.1.0"
description = ""
authors = ["Yves <rroughpatch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyobjc = "^10.3.1"
py-applescript = "^1.0.3"
pygetwindow = "^0.0.9"
psutil = "^6.0.0"
pypresence = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
